refactor: extend the id column in the hera_app_role table to the bigint type (#627)

+++ ozhera-app/app-api/src/main/java/org/apache/ozhera/app/api/service/HeraAppService.java
@@ -63,7 +63,7 @@ public interface HeraAppService {
 
     Long getAppCount();
 
-    Integer delRoleById(Integer id);
+    Integer delRoleById(Long id);
 
     Integer addRole(HeraAppRoleModel roleModel);
