+++ ozhera-app/app-service/src/main/java/org/apache/ozhera/app/service/impl/HeraAppServiceImpl.java
@@ -311,7 +311,7 @@ public class HeraAppServiceImpl implements HeraAppService {
     }
 
     @Override
-    public Integer delRoleById(Integer id) {
+    public Integer delRoleById(Long id) {
         return roleService.delById(id);
     }
